Revert "Implementation of hooks file for login and rearranged all feature fil…"
This reverts commit 90f745b8416e4eb936c48049203c6c3b060b5303.
