+++ src/main/resources/config.properties
@@ -2,7 +2,6 @@
  
 ## Base URL of the application
 baseUrl=https://workshop.cfg.deloitte.com/cfg-ai-demo/SemossWeb/packages/client/dist/#/login
-localUrl=http://localhost:9090/SemossWeb/packages/client/dist
 applicationName=GovConnect.AI
  
 ## Browser settings
@@ -18,7 +17,6 @@ ms_password=
 
 native_username=rramos@deloitte.com
 native_password=Test123!
-
 
 ## Test data file paths
 BIDataImportCSVFile=./src/test/resources/data/BIDataImportsTestData.csv
