+++ src/test/java/aicore/runners/AICoreTestRunner.java
@@ -6,8 +6,8 @@ import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(features = "src/test/resources/Features", glue = { "aicore.steps","aicore.hooks" }, plugin = { "pretty",
-		"html:target/cucumber-report.html" }, monochrome = true, dryRun = true )
+@CucumberOptions(features = "src/test/resources/Features", glue = { "aicore.steps" }, plugin = { "pretty",
+		"html:target/cucumber-report.html" }, monochrome = true, dryRun = true)
 
 public class AICoreTestRunner {
